fix(character-sheets): correct AuraData/AuraThemingData types to Decimal, tighten _build_magic_aura type
Fix TypedDict type contracts to use Decimal instead of int for aura
percentages, matching the actual DecimalField model. Tighten
_build_magic_aura parameter from Model to ObjectDB. Consolidate
Decimal imports to module level in tests.

+++ src/world/character_sheets/types.py
@@ -4,6 +4,7 @@ Type definitions for character sheets app.
 
 from __future__ import annotations
 
+from decimal import Decimal
 from typing import TypedDict
 
 from django.db import models
@@ -156,18 +157,18 @@ class AnimaRitualSection(TypedDict):
 class AuraData(TypedDict):
     """Full aura data including glimpse story (used in magic section)."""
 
-    celestial: int
-    primal: int
-    abyssal: int
+    celestial: Decimal
+    primal: Decimal
+    abyssal: Decimal
     glimpse_story: str
 
 
 class AuraThemingData(TypedDict):
     """Aura percentages for frontend styling (no glimpse story)."""
 
-    celestial: int
-    primal: int
-    abyssal: int
+    celestial: Decimal
+    primal: Decimal
+    abyssal: Decimal
 
 
 class MagicSection(TypedDict):
